--- src/main/java/med/voll/api/domain/usecases/output/IdOutput.java
@@ -1,4 +1,0 @@
-package med.voll.api.domain.usecases.output;
-
-public record IdOutput(String id) {
-}
