+++ src/main/java/med/voll/api/presentation/controllers/inputs/SpecialtyInput.java
@@ -1,4 +1,4 @@
-package med.voll.api.domain.usecases.inputs;
+package med.voll.api.presentation.controllers.inputs;
 
 public enum SpecialtyInput {
     ORTOPEDIA,
