+++ src/main/java/med/voll/api/presentation/controllers/inputs/address/AddressInput.java
@@ -1,6 +1,6 @@
-package med.voll.api.domain.usecases.inputs.address;
+package med.voll.api.presentation.controllers.inputs.address;
 
 public record AddressInput(String street, String neighborhood, String zip, String city, String state, String number,
-        String complement) {
+                String complement) {
 
 }
