+++ src/main/java/med/voll/api/presentation/controllers/inputs/doctor/SingUpInput.java
@@ -1,0 +1,7 @@
+package med.voll.api.presentation.controllers.inputs.doctor;
+
+import med.voll.api.presentation.controllers.inputs.address.AddressInput;
+import med.voll.api.presentation.controllers.inputs.SpecialtyInput;
+
+public record SingUpInput(String name, String email, String crm, SpecialtyInput specialty, AddressInput address) {
+}
