+++ src/main/java/med/voll/api/presentation/controllers/output/IdOutput.java
@@ -1,0 +1,4 @@
+package med.voll.api.presentation.controllers.output;
+
+public record IdOutput(String id) {
+}
